Major refactor of the initial crap. Better understanding of API.
Replacing child_process.spawn with atom's BufferedProcess which is
really just a wrapper for the spawn. Also changing the status command
to be a commit command since it will show the same info as status.
Adding some detail to package.json and making use of the File class as
well from pathwatcher module.