Fix bug #340 - commit editor/pane doesn't close.
In git.cmd, make the stderr callback identical to the stdout callback so that
they append to the same output string and make the exit callback reject(output)
if the exit code is non-zero, and resolve(output) if it is zero.

Fix git.getConfig to handle the rejection when a key is not set.

Add a catch() callback to GitCommit.commit that notifies the user of a commit
failure (so that failures won't be silent) and doesn't destroy the editor .